python-engine/requirements.txt
kafka-python==2.0.2
redis==5.0.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
sortedcontainers==2.4.0
ray==2.8.0
flask==3.0.0
pytest==7.4.3
tqdm==4.66.1
